Started some work on values from file but not working atm

+++ src/model/iOverlord.java
@@ -26,7 +26,7 @@ public interface iOverlord {
 	public boolean rotateGizmo(String gizmoName) throws CannotRotateException;
 	public void setGravity(float newGrav);
 	public void setFriction(float mu, float mu2);
-	public float getGravity();
+	public double getGravity();
 	public float getFrictionMu();
 	public float getFrictionMu2();
 	public boolean keyConnect(int keyNum, boolean direction, String consumer); //up = true, false = down.
